google_sheet2/read_google_sheet/requirements.txt
asgiref==3.5.2
cachetools==5.1.0
certifi==2022.5.18
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
google-api-core==2.8.0
google-api-python-client==2.48.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.1
greenlet==1.1.2
gspread==5.3.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.7
oauth2client==4.1.3
oauthlib==3.2.0
protobuf==3.20.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyTelegramBotAPI==4.5.1
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
